refactor(dashboard): move severity/cloud icons to assets and add --info theme token
- Move inline severity SVG components to assets/icons/ as standalone
  files imported via SVGR ?react, matching the established pattern
  used by RefreshIcon, PencilIcon, DiscordIcon, etc.
- Add severity_critical.svg, severity_warning.svg, severity_info.svg,
  cloud_done.svg with fill="currentColor"
- Remove cross-feature import from dtest into advisor/severityIcons
- Add --insforge-sky-500/600 channels and --info semantic token to
  packages/ui, register info: rgb(var(--info)) in dashboard tailwind
  config so all three severities (critical/warning/info) use theme
  tokens (text-destructive, text-warning, text-info) instead of
  hardcoded text-red-500/text-amber-500/text-sky-500
