--- packages/dashboard/src/features/dashboard/components/advisor/severityIcons.tsx
@@ -1,51 +1,0 @@
-interface SeverityIconProps {
-  className?: string;
-}
-
-export function CriticalIcon({ className }: SeverityIconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      className={className}
-      aria-hidden="true"
-    >
-      <path d="M9.3645 18.2292C9.16311 18.1459 8.97909 18.0278 8.81242 17.875L2.12492 11.1875C1.97214 11.0209 1.85409 10.8368 1.77075 10.6355C1.68742 10.4341 1.64575 10.2223 1.64575 10C1.64575 9.77782 1.68742 9.56254 1.77075 9.35421C1.85409 9.14587 1.97214 8.96532 2.12492 8.81254L8.81242 2.12504C8.97909 1.95837 9.16311 1.83685 9.3645 1.76046C9.56589 1.68407 9.7777 1.64587 9.99992 1.64587C10.2221 1.64587 10.4374 1.68407 10.6458 1.76046C10.8541 1.83685 11.0346 1.95837 11.1874 2.12504L17.8749 8.81254C18.0416 8.96532 18.1631 9.14587 18.2395 9.35421C18.3159 9.56254 18.3541 9.77782 18.3541 10C18.3541 10.2223 18.3159 10.4341 18.2395 10.6355C18.1631 10.8368 18.0416 11.0209 17.8749 11.1875L11.1874 17.875C11.0346 18.0278 10.8541 18.1459 10.6458 18.2292C10.4374 18.3125 10.2221 18.3542 9.99992 18.3542C9.7777 18.3542 9.56589 18.3125 9.3645 18.2292ZM9.99992 16.6875L16.6874 10L9.99992 3.31254L3.31242 10L9.99992 16.6875ZM9.16659 10.8334H10.8333V5.83337H9.16659V10.8334ZM10.5937 13.0938C10.7534 12.9341 10.8333 12.7362 10.8333 12.5C10.8333 12.2639 10.7534 12.066 10.5937 11.9063C10.4339 11.7466 10.236 11.6667 9.99992 11.6667C9.76381 11.6667 9.56589 11.7466 9.40617 11.9063C9.24645 12.066 9.16659 12.2639 9.16659 12.5C9.16659 12.7362 9.24645 12.9341 9.40617 13.0938C9.56589 13.2535 9.76381 13.3334 9.99992 13.3334C10.236 13.3334 10.4339 13.2535 10.5937 13.0938Z" />
-    </svg>
-  );
-}
-
-export function WarningIcon({ className }: SeverityIconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      className={className}
-      aria-hidden="true"
-    >
-      <path d="M9.99998 13.9423C10.1907 13.9423 10.3505 13.8778 10.4796 13.7488C10.6086 13.6198 10.6731 13.4599 10.6731 13.2692C10.6731 13.0785 10.6086 12.9187 10.4796 12.7896C10.3505 12.6607 10.1907 12.5963 9.99998 12.5963C9.80929 12.5963 9.64942 12.6607 9.5204 12.7896C9.39137 12.9187 9.32685 13.0785 9.32685 13.2692C9.32685 13.4599 9.39137 13.6198 9.5204 13.7488C9.64942 13.8778 9.80929 13.9423 9.99998 13.9423ZM9.37498 10.8975H10.625V5.89754H9.37498V10.8975ZM10.0014 17.9167C8.90644 17.9167 7.8772 17.7089 6.91373 17.2934C5.95026 16.8778 5.1122 16.3139 4.39956 15.6015C3.68692 14.8891 3.12269 14.0514 2.70685 13.0884C2.29116 12.1253 2.08331 11.0964 2.08331 10.0015C2.08331 8.9065 2.29109 7.87726 2.70665 6.91379C3.1222 5.95032 3.68616 5.11226 4.39852 4.39962C5.11088 3.68698 5.94859 3.12275 6.91165 2.70692C7.8747 2.29122 8.90366 2.08337 9.99852 2.08337C11.0935 2.08337 12.1228 2.29115 13.0862 2.70671C14.0497 3.12226 14.8878 3.68622 15.6004 4.39858C16.313 5.11094 16.8773 5.94865 17.2931 6.91171C17.7088 7.87476 17.9166 8.90372 17.9166 9.99858C17.9166 11.0936 17.7089 12.1228 17.2933 13.0863C16.8778 14.0498 16.3138 14.8878 15.6014 15.6005C14.8891 16.3131 14.0514 16.8773 13.0883 17.2932C12.1253 17.7089 11.0963 17.9167 10.0014 17.9167ZM9.99998 16.6667C11.8611 16.6667 13.4375 16.0209 14.7291 14.7292C16.0208 13.4375 16.6666 11.8612 16.6666 10C16.6666 8.13893 16.0208 6.56254 14.7291 5.27087C13.4375 3.97921 11.8611 3.33337 9.99998 3.33337C8.13887 3.33337 6.56248 3.97921 5.27081 5.27087C3.97915 6.56254 3.33331 8.13893 3.33331 10C3.33331 11.8612 3.97915 13.4375 5.27081 14.7292C6.56248 16.0209 8.13887 16.6667 9.99998 16.6667Z" />
-    </svg>
-  );
-}
-
-export function InfoIcon({ className }: SeverityIconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      className={className}
-      aria-hidden="true"
-    >
-      <path d="M9.99998 14.1667C10.1907 14.1667 10.3505 14.1023 10.4796 13.9733C10.6086 13.8442 10.6731 13.6843 10.6731 13.4936V9.66029C10.6731 9.4696 10.6086 9.30971 10.4796 9.18062C10.3505 9.05154 10.1907 8.987 9.99998 8.987C9.80929 8.987 9.64942 9.05154 9.5204 9.18062C9.39137 9.30971 9.32685 9.4696 9.32685 9.66029V13.4936C9.32685 13.6843 9.39137 13.8442 9.5204 13.9733C9.64942 14.1023 9.80929 14.1667 9.99998 14.1667ZM9.99998 7.40385C10.1907 7.40385 10.3505 7.33933 10.4796 7.21029C10.6086 7.08125 10.6731 6.92141 10.6731 6.73075C10.6731 6.54009 10.6086 6.38022 10.4796 6.25117C10.3505 6.12211 10.1907 6.05758 9.99998 6.05758C9.80929 6.05758 9.64942 6.12211 9.5204 6.25117C9.39137 6.38022 9.32685 6.54009 9.32685 6.73075C9.32685 6.92141 9.39137 7.08125 9.5204 7.21029C9.64942 7.33933 9.80929 7.40385 9.99998 7.40385ZM10.0014 17.9167C8.90644 17.9167 7.8772 17.7089 6.91373 17.2934C5.95026 16.8778 5.1122 16.3139 4.39956 15.6015C3.68692 14.8891 3.12269 14.0514 2.70685 13.0884C2.29116 12.1253 2.08331 11.0964 2.08331 10.0015C2.08331 8.9065 2.29109 7.87726 2.70665 6.91379C3.1222 5.95032 3.68616 5.11226 4.39852 4.39962C5.11088 3.68698 5.94859 3.12275 6.91165 2.70692C7.8747 2.29122 8.90366 2.08337 9.99852 2.08337C11.0935 2.08337 12.1228 2.29115 13.0862 2.70671C14.0497 3.12226 14.8878 3.68622 15.6004 4.39858C16.313 5.11094 16.8773 5.94865 17.2931 6.91171C17.7088 7.87476 17.9166 8.90372 17.9166 9.99858C17.9166 11.0936 17.7089 12.1228 17.2933 13.0863C16.8778 14.0498 16.3138 14.8878 15.6014 15.6005C14.8891 16.3131 14.0514 16.8773 13.0883 17.2932C12.1253 17.7089 11.0963 17.9167 10.0014 17.9167ZM9.99998 16.6667C11.8611 16.6667 13.4375 16.0209 14.7291 14.7292C16.0208 13.4375 16.6666 11.8612 16.6666 10C16.6666 8.13893 16.0208 6.56254 14.7291 5.27087C13.4375 3.97921 11.8611 3.33337 9.99998 3.33337C8.13887 3.33337 6.56248 3.97921 5.27081 5.27087C3.97915 6.56254 3.33331 8.13893 3.33331 10C3.33331 11.8612 3.97915 13.4375 5.27081 14.7292C6.56248 16.0209 8.13887 16.6667 9.99998 16.6667Z" />
-    </svg>
-  );
-}
-
-export const SEVERITY_ICONS = {
-  critical: CriticalIcon,
-  warning: WarningIcon,
-  info: InfoIcon,
-} as const;
